+++ bluez-async/src/adapter.rs
@@ -1,5 +1,8 @@
+use bluez_generated::OrgBluezAdapter1Properties;
 use dbus::Path;
 use std::fmt::{self, Display, Formatter};
+
+use crate::{AddressType, BluetoothError, MacAddress};
 
 /// Opaque identifier for a Bluetooth adapter on the system.
 #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
@@ -24,6 +27,88 @@ impl Display for AdapterId {
                 .to_string()
                 .strip_prefix("/org/bluez/")
                 .ok_or(fmt::Error)?
+        )
+    }
+}
+
+/// Information about a Bluetooth adapter on the system.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct AdapterInfo {
+    /// An opaque identifier for the adapter. This can be used to perform operations on it.
+    pub id: AdapterId,
+    /// The MAC address of the adapter.
+    pub mac_address: MacAddress,
+    /// The type of MAC address the adapter uses.
+    pub address_type: AddressType,
+    /// Whether the adapter is currently turned on.
+    pub powered: bool,
+    /// Whether the adapter is currently discovering devices.
+    pub discovering: bool,
+}
+
+impl AdapterInfo {
+    pub(crate) fn from_properties(
+        id: AdapterId,
+        adapter_properties: OrgBluezAdapter1Properties,
+    ) -> Result<AdapterInfo, BluetoothError> {
+        let mac_address = adapter_properties
+            .address()
+            .ok_or_else(|| BluetoothError::RequiredPropertyMissing("Address".to_string()))?;
+        let address_type = adapter_properties
+            .address_type()
+            .ok_or_else(|| BluetoothError::RequiredPropertyMissing("AddressType".to_string()))?
+            .parse()?;
+
+        Ok(AdapterInfo {
+            id,
+            mac_address: MacAddress(mac_address.to_owned()),
+            address_type,
+            powered: adapter_properties
+                .powered()
+                .ok_or_else(|| BluetoothError::RequiredPropertyMissing("Powered".to_string()))?,
+            discovering: adapter_properties.discovering().ok_or_else(|| {
+                BluetoothError::RequiredPropertyMissing("Discovering".to_string())
+            })?,
+        })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use dbus::arg::{RefArg, Variant};
+    use std::collections::HashMap;
+
+    use super::*;
+
+    #[test]
+    fn adapter_info_minimal() {
+        let id = AdapterId::new("/org/bluez/hci0");
+        let mut adapter_properties: HashMap<String, Variant<Box<dyn RefArg>>> = HashMap::new();
+        adapter_properties.insert(
+            "Address".to_string(),
+            Variant(Box::new("00:11:22:33:44:55".to_string())),
+        );
+        adapter_properties.insert(
+            "AddressType".to_string(),
+            Variant(Box::new("public".to_string())),
+        );
+        adapter_properties.insert("Powered".to_string(), Variant(Box::new(false)));
+        adapter_properties.insert("Discovering".to_string(), Variant(Box::new(false)));
+
+        let adapter = AdapterInfo::from_properties(
+            id.clone(),
+            OrgBluezAdapter1Properties(&adapter_properties),
+        )
+        .unwrap();
+        assert_eq!(
+            adapter,
+            AdapterInfo {
+                id,
+                mac_address: MacAddress("00:11:22:33:44:55".to_string()),
+                address_type: AddressType::Public,
+                powered: false,
+                discovering: false
+            }
         )
     }
 }